Move initial validation.js into old/
